fix some relations, update mapper profile, add DTOs and controllers for Ability and Specie

+++ PokemonLite.Domain/Entities/Pokemon.cs
@@ -5,7 +5,7 @@ public class Pokemon : BaseEntity
     public string Name { get; set; } = null!;
     public string Description { get; set; } = string.Empty;
     public ICollection<Specie> Species { get; set; } = [];
-    public double BaseHp { get; set; }
+    public int BaseHp { get; set; }
     public int BaseAttack { get; set; }
     public int BaseDefense { get; set; }
